src/requirements.txt
flask
mysql-connector-python
pandas
configparser
flask-sqlalchemy
mysqlclient